+++ docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+Version 1.0.1
+-------------
+
+* Rename apply() to compile()
+* add setTemplate() and setContent()
+* add more examples
+
 Version 1.0.0
 -------------